@@ -1,11 +1,12 @@
 module github.com/bot-games/battleships
 
-go 1.16
+go 1.18
 
 require (
-	github.com/bot-games/game-manager v0.4.1
-	github.com/bot-games/localrunner v0.1.0
-	github.com/go-qbit/rpc v0.0.0-20210926071738-84f99eae8c99
-	github.com/golang/protobuf v1.5.2
-	google.golang.org/protobuf v1.26.0
+	github.com/bot-games/game-manager v0.5.0
+	github.com/bot-games/localrunner v0.2.0
+	github.com/go-qbit/rpc v0.4.1
+	google.golang.org/protobuf v1.30.0
 )
+
+require github.com/google/uuid v1.3.0 // indirect
