@@ -1,11 +1,14 @@
 //go:generate gostatic2lib -path docs/ -package docs -out docs/game.go
+//go:generate gostatic2lib -path ../player/dist -package player -out ./player/dist.go
+
 package api
 
 import (
 	"bytes"
 	"compress/gzip"
 	"context"
-	"io/ioutil"
+	"io"
+	"net/http"
 
 	"github.com/go-qbit/rpc"
 	"github.com/go-qbit/rpc/openapi"
@@ -16,7 +19,7 @@ import (
 	mActionSkip "github.com/bot-games/battleships/api/method/action/skip"
 	mJoin "github.com/bot-games/battleships/api/method/join"
 	mWaitTurn "github.com/bot-games/battleships/api/method/wait_turn"
-
+	"github.com/bot-games/battleships/api/player"
 	"github.com/bot-games/game-manager"
 )
 
@@ -25,7 +28,7 @@ type BattleshipsRpc struct {
 }
 
 func New(gm *manager.GameManager) *BattleshipsRpc {
-	gameRpc := &BattleshipsRpc{rpc.New("github.com/bot-games/battleships/api/method")}
+	gameRpc := &BattleshipsRpc{rpc.New("github.com/bot-games/battleships/api/method", rpc.WithCors("*"))}
 
 	if err := gameRpc.RegisterMethods(
 		mJoin.New(gm),
@@ -45,9 +48,13 @@ func (r *BattleshipsRpc) GetSwagger(ctx context.Context) *openapi.OpenApi {
 	swagger.Info.Title = "Battleships bot API"
 
 	gz, _ := gzip.NewReader(bytes.NewBuffer(docs.NewHTTPHandler().GetFile("/game.md").Data))
-	data, _ := ioutil.ReadAll(gz)
+	data, _ := io.ReadAll(gz)
 
 	swagger.Info.Description = string(data)
 
 	return swagger
+}
+
+func (r *BattleshipsRpc) GetPlayerHandler() http.Handler {
+	return player.NewHTTPHandler()
 }
