@@ -1,7 +1,7 @@
 package battleships
 
 import (
-	"github.com/golang/protobuf/proto"
+	"google.golang.org/protobuf/proto"
 
 	"github.com/bot-games/battleships/pb"
 	"github.com/bot-games/game-manager"
