@@ -12,7 +12,7 @@ import (
 )
 
 var (
-	token  = flag.String("token", "", "")
+	token  = flag.String("token", "1", "")
 	gameId = flag.String("game", "", "")
 	debug  = flag.Bool("debug", false, "")
 )
@@ -25,7 +25,9 @@ func main() {
 		log.Fatalf("Empty token")
 	}
 
-	bsApi := api.New("https://bot-games.fun/battleships/api")
+	//bsApi := api.New("https://bot-games.fun/battleships/api")
+	//bsApi := api.New("http://localhost:10000/game")
+	bsApi := api.New("http://localhost:8080/game/battleships")
 
 	if *gameId == "" {
 		log.Println("Waiting opponent")
