@@ -5,15 +5,22 @@ import (
 	"github.com/bot-games/battleships/api"
 	manager "github.com/bot-games/game-manager"
 	"github.com/bot-games/localrunner"
+	"github.com/bot-games/localrunner/scheduler"
+	"github.com/bot-games/localrunner/storage"
 )
 
 func main() {
+	gameStorage := storage.New()
+
 	localrunner.Start(
-		api.New(
-			manager.New(battleships.Battleships{},
-				&localrunner.Storage{},
-				&localrunner.Scheduler{},
-			),
+		manager.New(
+			"battleships", "Battleships",
+			battleships.Battleships{},
+			gameStorage, scheduler.New(),
+			func(m *manager.GameManager) manager.GameApi {
+				return api.New(m)
+			},
 		),
+		gameStorage,
 	)
 }
