@@ -1,13 +1,12 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.25.0
-// 	protoc        v3.6.1
+// 	protoc-gen-go v1.30.0
+// 	protoc        v3.12.4
 // source: pb/state.proto
 
 package pb
 
 import (
-	proto "github.com/golang/protobuf/proto"
 	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
 	reflect "reflect"
@@ -20,10 +19,6 @@ const (
 	// Verify that runtime/protoimpl is sufficiently up-to-date.
 	_ = protoimpl.EnforceVersion(protoimpl.MaxVersion - 20)
 )
-
-// This is a compile-time assertion that a sufficiently up-to-date version
-// of the legacy proto package is being used.
-const _ = proto.ProtoPackageIsVersion4
 
 type Cell int32
 
@@ -77,6 +72,44 @@ func (Cell) EnumDescriptor() ([]byte, []int) {
 	return file_pb_state_proto_rawDescGZIP(), []int{0}
 }
 
+type Options struct {
+	state         protoimpl.MessageState
+	sizeCache     protoimpl.SizeCache
+	unknownFields protoimpl.UnknownFields
+}
+
+func (x *Options) Reset() {
+	*x = Options{}
+	if protoimpl.UnsafeEnabled {
+		mi := &file_pb_state_proto_msgTypes[0]
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		ms.StoreMessageInfo(mi)
+	}
+}
+
+func (x *Options) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*Options) ProtoMessage() {}
+
+func (x *Options) ProtoReflect() protoreflect.Message {
+	mi := &file_pb_state_proto_msgTypes[0]
+	if protoimpl.UnsafeEnabled && x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use Options.ProtoReflect.Descriptor instead.
+func (*Options) Descriptor() ([]byte, []int) {
+	return file_pb_state_proto_rawDescGZIP(), []int{0}
+}
+
 type State struct {
 	state         protoimpl.MessageState
 	sizeCache     protoimpl.SizeCache
@@ -89,7 +122,7 @@ type State struct {
 func (x *State) Reset() {
 	*x = State{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_pb_state_proto_msgTypes[0]
+		mi := &file_pb_state_proto_msgTypes[1]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -102,7 +135,7 @@ func (x *State) String() string {
 func (*State) ProtoMessage() {}
 
 func (x *State) ProtoReflect() protoreflect.Message {
-	mi := &file_pb_state_proto_msgTypes[0]
+	mi := &file_pb_state_proto_msgTypes[1]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -115,7 +148,7 @@ func (x *State) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use State.ProtoReflect.Descriptor instead.
 func (*State) Descriptor() ([]byte, []int) {
-	return file_pb_state_proto_rawDescGZIP(), []int{0}
+	return file_pb_state_proto_rawDescGZIP(), []int{1}
 }
 
 func (x *State) GetField1() []Cell {
@@ -138,6 +171,7 @@ type Action struct {
 	unknownFields protoimpl.UnknownFields
 
 	// Types that are assignable to Data:
+	//
 	//	*Action_Skip
 	//	*Action_Setup
 	//	*Action_Fire
@@ -147,7 +181,7 @@ type Action struct {
 func (x *Action) Reset() {
 	*x = Action{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_pb_state_proto_msgTypes[1]
+		mi := &file_pb_state_proto_msgTypes[2]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -160,7 +194,7 @@ func (x *Action) String() string {
 func (*Action) ProtoMessage() {}
 
 func (x *Action) ProtoReflect() protoreflect.Message {
-	mi := &file_pb_state_proto_msgTypes[1]
+	mi := &file_pb_state_proto_msgTypes[2]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -173,7 +207,7 @@ func (x *Action) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use Action.ProtoReflect.Descriptor instead.
 func (*Action) Descriptor() ([]byte, []int) {
-	return file_pb_state_proto_rawDescGZIP(), []int{1}
+	return file_pb_state_proto_rawDescGZIP(), []int{2}
 }
 
 func (m *Action) GetData() isAction_Data {
@@ -235,7 +269,7 @@ type ActionSkip struct {
 func (x *ActionSkip) Reset() {
 	*x = ActionSkip{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_pb_state_proto_msgTypes[2]
+		mi := &file_pb_state_proto_msgTypes[3]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -248,7 +282,7 @@ func (x *ActionSkip) String() string {
 func (*ActionSkip) ProtoMessage() {}
 
 func (x *ActionSkip) ProtoReflect() protoreflect.Message {
-	mi := &file_pb_state_proto_msgTypes[2]
+	mi := &file_pb_state_proto_msgTypes[3]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -261,7 +295,7 @@ func (x *ActionSkip) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use ActionSkip.ProtoReflect.Descriptor instead.
 func (*ActionSkip) Descriptor() ([]byte, []int) {
-	return file_pb_state_proto_rawDescGZIP(), []int{2}
+	return file_pb_state_proto_rawDescGZIP(), []int{3}
 }
 
 type ActionSetup struct {
@@ -284,7 +318,7 @@ type ActionSetup struct {
 func (x *ActionSetup) Reset() {
 	*x = ActionSetup{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_pb_state_proto_msgTypes[3]
+		mi := &file_pb_state_proto_msgTypes[4]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -297,7 +331,7 @@ func (x *ActionSetup) String() string {
 func (*ActionSetup) ProtoMessage() {}
 
 func (x *ActionSetup) ProtoReflect() protoreflect.Message {
-	mi := &file_pb_state_proto_msgTypes[3]
+	mi := &file_pb_state_proto_msgTypes[4]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -310,7 +344,7 @@ func (x *ActionSetup) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use ActionSetup.ProtoReflect.Descriptor instead.
 func (*ActionSetup) Descriptor() ([]byte, []int) {
-	return file_pb_state_proto_rawDescGZIP(), []int{3}
+	return file_pb_state_proto_rawDescGZIP(), []int{4}
 }
 
 func (x *ActionSetup) GetShipL4N1() *ActionSetup_Ship {
@@ -394,7 +428,7 @@ type ActionFire struct {
 func (x *ActionFire) Reset() {
 	*x = ActionFire{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_pb_state_proto_msgTypes[4]
+		mi := &file_pb_state_proto_msgTypes[5]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -407,7 +441,7 @@ func (x *ActionFire) String() string {
 func (*ActionFire) ProtoMessage() {}
 
 func (x *ActionFire) ProtoReflect() protoreflect.Message {
-	mi := &file_pb_state_proto_msgTypes[4]
+	mi := &file_pb_state_proto_msgTypes[5]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -420,7 +454,7 @@ func (x *ActionFire) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use ActionFire.ProtoReflect.Descriptor instead.
 func (*ActionFire) Descriptor() ([]byte, []int) {
-	return file_pb_state_proto_rawDescGZIP(), []int{4}
+	return file_pb_state_proto_rawDescGZIP(), []int{5}
 }
 
 func (x *ActionFire) GetCoordinate() string {
@@ -442,7 +476,7 @@ type ActionSetup_Ship struct {
 func (x *ActionSetup_Ship) Reset() {
 	*x = ActionSetup_Ship{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_pb_state_proto_msgTypes[5]
+		mi := &file_pb_state_proto_msgTypes[6]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -455,7 +489,7 @@ func (x *ActionSetup_Ship) String() string {
 func (*ActionSetup_Ship) ProtoMessage() {}
 
 func (x *ActionSetup_Ship) ProtoReflect() protoreflect.Message {
-	mi := &file_pb_state_proto_msgTypes[5]
+	mi := &file_pb_state_proto_msgTypes[6]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -468,7 +502,7 @@ func (x *ActionSetup_Ship) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use ActionSetup_Ship.ProtoReflect.Descriptor instead.
 func (*ActionSetup_Ship) Descriptor() ([]byte, []int) {
-	return file_pb_state_proto_rawDescGZIP(), []int{3, 0}
+	return file_pb_state_proto_rawDescGZIP(), []int{4, 0}
 }
 
 func (x *ActionSetup_Ship) GetCoordinate() string {
@@ -489,61 +523,62 @@ var File_pb_state_proto protoreflect.FileDescriptor
 
 var file_pb_state_proto_rawDesc = []byte{
 	0x0a, 0x0e, 0x70, 0x62, 0x2f, 0x73, 0x74, 0x61, 0x74, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f,
-	0x22, 0x45, 0x0a, 0x05, 0x53, 0x74, 0x61, 0x74, 0x65, 0x12, 0x1d, 0x0a, 0x06, 0x66, 0x69, 0x65,
-	0x6c, 0x64, 0x31, 0x18, 0x01, 0x20, 0x03, 0x28, 0x0e, 0x32, 0x05, 0x2e, 0x43, 0x65, 0x6c, 0x6c,
-	0x52, 0x06, 0x66, 0x69, 0x65, 0x6c, 0x64, 0x31, 0x12, 0x1d, 0x0a, 0x06, 0x66, 0x69, 0x65, 0x6c,
-	0x64, 0x32, 0x18, 0x02, 0x20, 0x03, 0x28, 0x0e, 0x32, 0x05, 0x2e, 0x43, 0x65, 0x6c, 0x6c, 0x52,
-	0x06, 0x66, 0x69, 0x65, 0x6c, 0x64, 0x32, 0x22, 0x7c, 0x0a, 0x06, 0x41, 0x63, 0x74, 0x69, 0x6f,
-	0x6e, 0x12, 0x21, 0x0a, 0x04, 0x73, 0x6b, 0x69, 0x70, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0b, 0x32,
-	0x0b, 0x2e, 0x41, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x53, 0x6b, 0x69, 0x70, 0x48, 0x00, 0x52, 0x04,
-	0x73, 0x6b, 0x69, 0x70, 0x12, 0x24, 0x0a, 0x05, 0x73, 0x65, 0x74, 0x75, 0x70, 0x18, 0x02, 0x20,
-	0x01, 0x28, 0x0b, 0x32, 0x0c, 0x2e, 0x41, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x53, 0x65, 0x74, 0x75,
-	0x70, 0x48, 0x00, 0x52, 0x05, 0x73, 0x65, 0x74, 0x75, 0x70, 0x12, 0x21, 0x0a, 0x04, 0x66, 0x69,
-	0x72, 0x65, 0x18, 0x03, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x0b, 0x2e, 0x41, 0x63, 0x74, 0x69, 0x6f,
-	0x6e, 0x46, 0x69, 0x72, 0x65, 0x48, 0x00, 0x52, 0x04, 0x66, 0x69, 0x72, 0x65, 0x42, 0x06, 0x0a,
-	0x04, 0x64, 0x61, 0x74, 0x61, 0x22, 0x0c, 0x0a, 0x0a, 0x41, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x53,
-	0x6b, 0x69, 0x70, 0x22, 0xa7, 0x04, 0x0a, 0x0b, 0x41, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x53, 0x65,
-	0x74, 0x75, 0x70, 0x12, 0x2d, 0x0a, 0x08, 0x73, 0x68, 0x69, 0x70, 0x4c, 0x34, 0x4e, 0x31, 0x18,
-	0x01, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x11, 0x2e, 0x41, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x53, 0x65,
-	0x74, 0x75, 0x70, 0x2e, 0x53, 0x68, 0x69, 0x70, 0x52, 0x08, 0x73, 0x68, 0x69, 0x70, 0x4c, 0x34,
-	0x4e, 0x31, 0x12, 0x2d, 0x0a, 0x08, 0x73, 0x68, 0x69, 0x70, 0x4c, 0x33, 0x4e, 0x31, 0x18, 0x02,
-	0x20, 0x01, 0x28, 0x0b, 0x32, 0x11, 0x2e, 0x41, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x53, 0x65, 0x74,
-	0x75, 0x70, 0x2e, 0x53, 0x68, 0x69, 0x70, 0x52, 0x08, 0x73, 0x68, 0x69, 0x70, 0x4c, 0x33, 0x4e,
-	0x31, 0x12, 0x2d, 0x0a, 0x08, 0x73, 0x68, 0x69, 0x70, 0x4c, 0x33, 0x4e, 0x32, 0x18, 0x03, 0x20,
-	0x01, 0x28, 0x0b, 0x32, 0x11, 0x2e, 0x41, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x53, 0x65, 0x74, 0x75,
-	0x70, 0x2e, 0x53, 0x68, 0x69, 0x70, 0x52, 0x08, 0x73, 0x68, 0x69, 0x70, 0x4c, 0x33, 0x4e, 0x32,
-	0x12, 0x2d, 0x0a, 0x08, 0x73, 0x68, 0x69, 0x70, 0x4c, 0x32, 0x4e, 0x31, 0x18, 0x04, 0x20, 0x01,
-	0x28, 0x0b, 0x32, 0x11, 0x2e, 0x41, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x53, 0x65, 0x74, 0x75, 0x70,
-	0x2e, 0x53, 0x68, 0x69, 0x70, 0x52, 0x08, 0x73, 0x68, 0x69, 0x70, 0x4c, 0x32, 0x4e, 0x31, 0x12,
-	0x2d, 0x0a, 0x08, 0x73, 0x68, 0x69, 0x70, 0x4c, 0x32, 0x4e, 0x32, 0x18, 0x05, 0x20, 0x01, 0x28,
-	0x0b, 0x32, 0x11, 0x2e, 0x41, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x53, 0x65, 0x74, 0x75, 0x70, 0x2e,
-	0x53, 0x68, 0x69, 0x70, 0x52, 0x08, 0x73, 0x68, 0x69, 0x70, 0x4c, 0x32, 0x4e, 0x32, 0x12, 0x2d,
-	0x0a, 0x08, 0x73, 0x68, 0x69, 0x70, 0x4c, 0x32, 0x4e, 0x33, 0x18, 0x06, 0x20, 0x01, 0x28, 0x0b,
+	0x22, 0x09, 0x0a, 0x07, 0x4f, 0x70, 0x74, 0x69, 0x6f, 0x6e, 0x73, 0x22, 0x45, 0x0a, 0x05, 0x53,
+	0x74, 0x61, 0x74, 0x65, 0x12, 0x1d, 0x0a, 0x06, 0x66, 0x69, 0x65, 0x6c, 0x64, 0x31, 0x18, 0x01,
+	0x20, 0x03, 0x28, 0x0e, 0x32, 0x05, 0x2e, 0x43, 0x65, 0x6c, 0x6c, 0x52, 0x06, 0x66, 0x69, 0x65,
+	0x6c, 0x64, 0x31, 0x12, 0x1d, 0x0a, 0x06, 0x66, 0x69, 0x65, 0x6c, 0x64, 0x32, 0x18, 0x02, 0x20,
+	0x03, 0x28, 0x0e, 0x32, 0x05, 0x2e, 0x43, 0x65, 0x6c, 0x6c, 0x52, 0x06, 0x66, 0x69, 0x65, 0x6c,
+	0x64, 0x32, 0x22, 0x7c, 0x0a, 0x06, 0x41, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x12, 0x21, 0x0a, 0x04,
+	0x73, 0x6b, 0x69, 0x70, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x0b, 0x2e, 0x41, 0x63, 0x74,
+	0x69, 0x6f, 0x6e, 0x53, 0x6b, 0x69, 0x70, 0x48, 0x00, 0x52, 0x04, 0x73, 0x6b, 0x69, 0x70, 0x12,
+	0x24, 0x0a, 0x05, 0x73, 0x65, 0x74, 0x75, 0x70, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x0c,
+	0x2e, 0x41, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x53, 0x65, 0x74, 0x75, 0x70, 0x48, 0x00, 0x52, 0x05,
+	0x73, 0x65, 0x74, 0x75, 0x70, 0x12, 0x21, 0x0a, 0x04, 0x66, 0x69, 0x72, 0x65, 0x18, 0x03, 0x20,
+	0x01, 0x28, 0x0b, 0x32, 0x0b, 0x2e, 0x41, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x46, 0x69, 0x72, 0x65,
+	0x48, 0x00, 0x52, 0x04, 0x66, 0x69, 0x72, 0x65, 0x42, 0x06, 0x0a, 0x04, 0x64, 0x61, 0x74, 0x61,
+	0x22, 0x0c, 0x0a, 0x0a, 0x41, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x53, 0x6b, 0x69, 0x70, 0x22, 0xa7,
+	0x04, 0x0a, 0x0b, 0x41, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x53, 0x65, 0x74, 0x75, 0x70, 0x12, 0x2d,
+	0x0a, 0x08, 0x73, 0x68, 0x69, 0x70, 0x4c, 0x34, 0x4e, 0x31, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0b,
 	0x32, 0x11, 0x2e, 0x41, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x53, 0x65, 0x74, 0x75, 0x70, 0x2e, 0x53,
-	0x68, 0x69, 0x70, 0x52, 0x08, 0x73, 0x68, 0x69, 0x70, 0x4c, 0x32, 0x4e, 0x33, 0x12, 0x2d, 0x0a,
-	0x08, 0x73, 0x68, 0x69, 0x70, 0x4c, 0x31, 0x4e, 0x31, 0x18, 0x07, 0x20, 0x01, 0x28, 0x0b, 0x32,
+	0x68, 0x69, 0x70, 0x52, 0x08, 0x73, 0x68, 0x69, 0x70, 0x4c, 0x34, 0x4e, 0x31, 0x12, 0x2d, 0x0a,
+	0x08, 0x73, 0x68, 0x69, 0x70, 0x4c, 0x33, 0x4e, 0x31, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0b, 0x32,
 	0x11, 0x2e, 0x41, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x53, 0x65, 0x74, 0x75, 0x70, 0x2e, 0x53, 0x68,
-	0x69, 0x70, 0x52, 0x08, 0x73, 0x68, 0x69, 0x70, 0x4c, 0x31, 0x4e, 0x31, 0x12, 0x2d, 0x0a, 0x08,
-	0x73, 0x68, 0x69, 0x70, 0x4c, 0x31, 0x4e, 0x32, 0x18, 0x08, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x11,
+	0x69, 0x70, 0x52, 0x08, 0x73, 0x68, 0x69, 0x70, 0x4c, 0x33, 0x4e, 0x31, 0x12, 0x2d, 0x0a, 0x08,
+	0x73, 0x68, 0x69, 0x70, 0x4c, 0x33, 0x4e, 0x32, 0x18, 0x03, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x11,
 	0x2e, 0x41, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x53, 0x65, 0x74, 0x75, 0x70, 0x2e, 0x53, 0x68, 0x69,
-	0x70, 0x52, 0x08, 0x73, 0x68, 0x69, 0x70, 0x4c, 0x31, 0x4e, 0x32, 0x12, 0x2d, 0x0a, 0x08, 0x73,
-	0x68, 0x69, 0x70, 0x4c, 0x31, 0x4e, 0x33, 0x18, 0x09, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x11, 0x2e,
+	0x70, 0x52, 0x08, 0x73, 0x68, 0x69, 0x70, 0x4c, 0x33, 0x4e, 0x32, 0x12, 0x2d, 0x0a, 0x08, 0x73,
+	0x68, 0x69, 0x70, 0x4c, 0x32, 0x4e, 0x31, 0x18, 0x04, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x11, 0x2e,
 	0x41, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x53, 0x65, 0x74, 0x75, 0x70, 0x2e, 0x53, 0x68, 0x69, 0x70,
-	0x52, 0x08, 0x73, 0x68, 0x69, 0x70, 0x4c, 0x31, 0x4e, 0x33, 0x12, 0x2d, 0x0a, 0x08, 0x73, 0x68,
-	0x69, 0x70, 0x4c, 0x31, 0x4e, 0x34, 0x18, 0x0a, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x11, 0x2e, 0x41,
+	0x52, 0x08, 0x73, 0x68, 0x69, 0x70, 0x4c, 0x32, 0x4e, 0x31, 0x12, 0x2d, 0x0a, 0x08, 0x73, 0x68,
+	0x69, 0x70, 0x4c, 0x32, 0x4e, 0x32, 0x18, 0x05, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x11, 0x2e, 0x41,
 	0x63, 0x74, 0x69, 0x6f, 0x6e, 0x53, 0x65, 0x74, 0x75, 0x70, 0x2e, 0x53, 0x68, 0x69, 0x70, 0x52,
-	0x08, 0x73, 0x68, 0x69, 0x70, 0x4c, 0x31, 0x4e, 0x34, 0x1a, 0x42, 0x0a, 0x04, 0x53, 0x68, 0x69,
-	0x70, 0x12, 0x1e, 0x0a, 0x0a, 0x63, 0x6f, 0x6f, 0x72, 0x64, 0x69, 0x6e, 0x61, 0x74, 0x65, 0x18,
-	0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x0a, 0x63, 0x6f, 0x6f, 0x72, 0x64, 0x69, 0x6e, 0x61, 0x74,
-	0x65, 0x12, 0x1a, 0x0a, 0x08, 0x76, 0x65, 0x72, 0x74, 0x69, 0x63, 0x61, 0x6c, 0x18, 0x02, 0x20,
-	0x01, 0x28, 0x08, 0x52, 0x08, 0x76, 0x65, 0x72, 0x74, 0x69, 0x63, 0x61, 0x6c, 0x22, 0x2c, 0x0a,
-	0x0a, 0x41, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x46, 0x69, 0x72, 0x65, 0x12, 0x1e, 0x0a, 0x0a, 0x63,
-	0x6f, 0x6f, 0x72, 0x64, 0x69, 0x6e, 0x61, 0x74, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52,
-	0x0a, 0x63, 0x6f, 0x6f, 0x72, 0x64, 0x69, 0x6e, 0x61, 0x74, 0x65, 0x2a, 0x30, 0x0a, 0x04, 0x43,
-	0x65, 0x6c, 0x6c, 0x12, 0x09, 0x0a, 0x05, 0x45, 0x4d, 0x50, 0x54, 0x59, 0x10, 0x00, 0x12, 0x08,
-	0x0a, 0x04, 0x53, 0x48, 0x49, 0x50, 0x10, 0x01, 0x12, 0x0a, 0x0a, 0x06, 0x4d, 0x49, 0x53, 0x53,
-	0x45, 0x44, 0x10, 0x02, 0x12, 0x07, 0x0a, 0x03, 0x47, 0x4f, 0x54, 0x10, 0x03, 0x42, 0x05, 0x5a,
-	0x03, 0x2f, 0x70, 0x62, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
+	0x08, 0x73, 0x68, 0x69, 0x70, 0x4c, 0x32, 0x4e, 0x32, 0x12, 0x2d, 0x0a, 0x08, 0x73, 0x68, 0x69,
+	0x70, 0x4c, 0x32, 0x4e, 0x33, 0x18, 0x06, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x11, 0x2e, 0x41, 0x63,
+	0x74, 0x69, 0x6f, 0x6e, 0x53, 0x65, 0x74, 0x75, 0x70, 0x2e, 0x53, 0x68, 0x69, 0x70, 0x52, 0x08,
+	0x73, 0x68, 0x69, 0x70, 0x4c, 0x32, 0x4e, 0x33, 0x12, 0x2d, 0x0a, 0x08, 0x73, 0x68, 0x69, 0x70,
+	0x4c, 0x31, 0x4e, 0x31, 0x18, 0x07, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x11, 0x2e, 0x41, 0x63, 0x74,
+	0x69, 0x6f, 0x6e, 0x53, 0x65, 0x74, 0x75, 0x70, 0x2e, 0x53, 0x68, 0x69, 0x70, 0x52, 0x08, 0x73,
+	0x68, 0x69, 0x70, 0x4c, 0x31, 0x4e, 0x31, 0x12, 0x2d, 0x0a, 0x08, 0x73, 0x68, 0x69, 0x70, 0x4c,
+	0x31, 0x4e, 0x32, 0x18, 0x08, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x11, 0x2e, 0x41, 0x63, 0x74, 0x69,
+	0x6f, 0x6e, 0x53, 0x65, 0x74, 0x75, 0x70, 0x2e, 0x53, 0x68, 0x69, 0x70, 0x52, 0x08, 0x73, 0x68,
+	0x69, 0x70, 0x4c, 0x31, 0x4e, 0x32, 0x12, 0x2d, 0x0a, 0x08, 0x73, 0x68, 0x69, 0x70, 0x4c, 0x31,
+	0x4e, 0x33, 0x18, 0x09, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x11, 0x2e, 0x41, 0x63, 0x74, 0x69, 0x6f,
+	0x6e, 0x53, 0x65, 0x74, 0x75, 0x70, 0x2e, 0x53, 0x68, 0x69, 0x70, 0x52, 0x08, 0x73, 0x68, 0x69,
+	0x70, 0x4c, 0x31, 0x4e, 0x33, 0x12, 0x2d, 0x0a, 0x08, 0x73, 0x68, 0x69, 0x70, 0x4c, 0x31, 0x4e,
+	0x34, 0x18, 0x0a, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x11, 0x2e, 0x41, 0x63, 0x74, 0x69, 0x6f, 0x6e,
+	0x53, 0x65, 0x74, 0x75, 0x70, 0x2e, 0x53, 0x68, 0x69, 0x70, 0x52, 0x08, 0x73, 0x68, 0x69, 0x70,
+	0x4c, 0x31, 0x4e, 0x34, 0x1a, 0x42, 0x0a, 0x04, 0x53, 0x68, 0x69, 0x70, 0x12, 0x1e, 0x0a, 0x0a,
+	0x63, 0x6f, 0x6f, 0x72, 0x64, 0x69, 0x6e, 0x61, 0x74, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09,
+	0x52, 0x0a, 0x63, 0x6f, 0x6f, 0x72, 0x64, 0x69, 0x6e, 0x61, 0x74, 0x65, 0x12, 0x1a, 0x0a, 0x08,
+	0x76, 0x65, 0x72, 0x74, 0x69, 0x63, 0x61, 0x6c, 0x18, 0x02, 0x20, 0x01, 0x28, 0x08, 0x52, 0x08,
+	0x76, 0x65, 0x72, 0x74, 0x69, 0x63, 0x61, 0x6c, 0x22, 0x2c, 0x0a, 0x0a, 0x41, 0x63, 0x74, 0x69,
+	0x6f, 0x6e, 0x46, 0x69, 0x72, 0x65, 0x12, 0x1e, 0x0a, 0x0a, 0x63, 0x6f, 0x6f, 0x72, 0x64, 0x69,
+	0x6e, 0x61, 0x74, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x0a, 0x63, 0x6f, 0x6f, 0x72,
+	0x64, 0x69, 0x6e, 0x61, 0x74, 0x65, 0x2a, 0x30, 0x0a, 0x04, 0x43, 0x65, 0x6c, 0x6c, 0x12, 0x09,
+	0x0a, 0x05, 0x45, 0x4d, 0x50, 0x54, 0x59, 0x10, 0x00, 0x12, 0x08, 0x0a, 0x04, 0x53, 0x48, 0x49,
+	0x50, 0x10, 0x01, 0x12, 0x0a, 0x0a, 0x06, 0x4d, 0x49, 0x53, 0x53, 0x45, 0x44, 0x10, 0x02, 0x12,
+	0x07, 0x0a, 0x03, 0x47, 0x4f, 0x54, 0x10, 0x03, 0x42, 0x05, 0x5a, 0x03, 0x2f, 0x70, 0x62, 0x62,
+	0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
 }
 
 var (
@@ -559,32 +594,33 @@ func file_pb_state_proto_rawDescGZIP() []byte {
 }
 
 var file_pb_state_proto_enumTypes = make([]protoimpl.EnumInfo, 1)
-var file_pb_state_proto_msgTypes = make([]protoimpl.MessageInfo, 6)
+var file_pb_state_proto_msgTypes = make([]protoimpl.MessageInfo, 7)
 var file_pb_state_proto_goTypes = []interface{}{
 	(Cell)(0),                // 0: Cell
-	(*State)(nil),            // 1: State
-	(*Action)(nil),           // 2: Action
-	(*ActionSkip)(nil),       // 3: ActionSkip
-	(*ActionSetup)(nil),      // 4: ActionSetup
-	(*ActionFire)(nil),       // 5: ActionFire
-	(*ActionSetup_Ship)(nil), // 6: ActionSetup.Ship
+	(*Options)(nil),          // 1: Options
+	(*State)(nil),            // 2: State
+	(*Action)(nil),           // 3: Action
+	(*ActionSkip)(nil),       // 4: ActionSkip
+	(*ActionSetup)(nil),      // 5: ActionSetup
+	(*ActionFire)(nil),       // 6: ActionFire
+	(*ActionSetup_Ship)(nil), // 7: ActionSetup.Ship
 }
 var file_pb_state_proto_depIdxs = []int32{
 	0,  // 0: State.field1:type_name -> Cell
 	0,  // 1: State.field2:type_name -> Cell
-	3,  // 2: Action.skip:type_name -> ActionSkip
-	4,  // 3: Action.setup:type_name -> ActionSetup
-	5,  // 4: Action.fire:type_name -> ActionFire
-	6,  // 5: ActionSetup.shipL4N1:type_name -> ActionSetup.Ship
-	6,  // 6: ActionSetup.shipL3N1:type_name -> ActionSetup.Ship
-	6,  // 7: ActionSetup.shipL3N2:type_name -> ActionSetup.Ship
-	6,  // 8: ActionSetup.shipL2N1:type_name -> ActionSetup.Ship
-	6,  // 9: ActionSetup.shipL2N2:type_name -> ActionSetup.Ship
-	6,  // 10: ActionSetup.shipL2N3:type_name -> ActionSetup.Ship
-	6,  // 11: ActionSetup.shipL1N1:type_name -> ActionSetup.Ship
-	6,  // 12: ActionSetup.shipL1N2:type_name -> ActionSetup.Ship
-	6,  // 13: ActionSetup.shipL1N3:type_name -> ActionSetup.Ship
-	6,  // 14: ActionSetup.shipL1N4:type_name -> ActionSetup.Ship
+	4,  // 2: Action.skip:type_name -> ActionSkip
+	5,  // 3: Action.setup:type_name -> ActionSetup
+	6,  // 4: Action.fire:type_name -> ActionFire
+	7,  // 5: ActionSetup.shipL4N1:type_name -> ActionSetup.Ship
+	7,  // 6: ActionSetup.shipL3N1:type_name -> ActionSetup.Ship
+	7,  // 7: ActionSetup.shipL3N2:type_name -> ActionSetup.Ship
+	7,  // 8: ActionSetup.shipL2N1:type_name -> ActionSetup.Ship
+	7,  // 9: ActionSetup.shipL2N2:type_name -> ActionSetup.Ship
+	7,  // 10: ActionSetup.shipL2N3:type_name -> ActionSetup.Ship
+	7,  // 11: ActionSetup.shipL1N1:type_name -> ActionSetup.Ship
+	7,  // 12: ActionSetup.shipL1N2:type_name -> ActionSetup.Ship
+	7,  // 13: ActionSetup.shipL1N3:type_name -> ActionSetup.Ship
+	7,  // 14: ActionSetup.shipL1N4:type_name -> ActionSetup.Ship
 	15, // [15:15] is the sub-list for method output_type
 	15, // [15:15] is the sub-list for method input_type
 	15, // [15:15] is the sub-list for extension type_name
@@ -599,7 +635,7 @@ func file_pb_state_proto_init() {
 	}
 	if !protoimpl.UnsafeEnabled {
 		file_pb_state_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*State); i {
+			switch v := v.(*Options); i {
 			case 0:
 				return &v.state
 			case 1:
@@ -611,7 +647,7 @@ func file_pb_state_proto_init() {
 			}
 		}
 		file_pb_state_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*Action); i {
+			switch v := v.(*State); i {
 			case 0:
 				return &v.state
 			case 1:
@@ -623,7 +659,7 @@ func file_pb_state_proto_init() {
 			}
 		}
 		file_pb_state_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*ActionSkip); i {
+			switch v := v.(*Action); i {
 			case 0:
 				return &v.state
 			case 1:
@@ -635,7 +671,7 @@ func file_pb_state_proto_init() {
 			}
 		}
 		file_pb_state_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*ActionSetup); i {
+			switch v := v.(*ActionSkip); i {
 			case 0:
 				return &v.state
 			case 1:
@@ -647,7 +683,7 @@ func file_pb_state_proto_init() {
 			}
 		}
 		file_pb_state_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*ActionFire); i {
+			switch v := v.(*ActionSetup); i {
 			case 0:
 				return &v.state
 			case 1:
@@ -659,6 +695,18 @@ func file_pb_state_proto_init() {
 			}
 		}
 		file_pb_state_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+			switch v := v.(*ActionFire); i {
+			case 0:
+				return &v.state
+			case 1:
+				return &v.sizeCache
+			case 2:
+				return &v.unknownFields
+			default:
+				return nil
+			}
+		}
+		file_pb_state_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
 			switch v := v.(*ActionSetup_Ship); i {
 			case 0:
 				return &v.state
@@ -671,7 +719,7 @@ func file_pb_state_proto_init() {
 			}
 		}
 	}
-	file_pb_state_proto_msgTypes[1].OneofWrappers = []interface{}{
+	file_pb_state_proto_msgTypes[2].OneofWrappers = []interface{}{
 		(*Action_Skip)(nil),
 		(*Action_Setup)(nil),
 		(*Action_Fire)(nil),
@@ -682,7 +730,7 @@ func file_pb_state_proto_init() {
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
 			RawDescriptor: file_pb_state_proto_rawDesc,
 			NumEnums:      1,
-			NumMessages:   6,
+			NumMessages:   7,
 			NumExtensions: 0,
 			NumServices:   0,
 		},
